convert account update to plain nodejs to use internal Test module
